feat(sooth_book): W4 probability-WAD price ladder
Translated price-bearing state, instruction arguments, events, trades, order requests, matching pools/queues, market liquidities, market outcomes, price ladders, and order accounts from f64 decimal odds to u128 probability-WAD.

Updated price math in instructions/math.rs and market-position/order paths: YES cost is stake * price / WAD, NO risk is stake * (WAD - price) / WAD, matched payout is one unit per share, and PDA price seeds now use u128 little-endian bytes.

Replaced DEFAULT_PRICES with generated 0.001-WAD ticks from 0.001 through 0.999 WAD (999 non-terminal buckets); validation rejects 0/WAD endpoints and non-tick prices.

Updated fixtures across matching, cancellation, settlement, liquidity, and price-ladder tests, including a decimal-odds conversion check for 2.5 -> 0.4 WAD. Regenerated and synced packages/sdk-solana/src/anchor/sooth_book.json from Anchor target IDL; price fields/vectors are u128 and no f64 remains.

Diff: 33 files changed, 833 insertions(+), 1700 deletions(-). Tests: cargo build -p sooth_book; cargo test -p sooth_book (205 -> 207 passing); anchor build.

+++ packages/programs-core/programs/sooth_book/src/events/trade.rs
@@ -3,6 +3,6 @@ use anchor_lang::prelude::*;
 #[event]
 pub struct TradeEvent {
     pub amount: u64,
-    pub price: f64,
+    pub price: u128,
     pub market: Pubkey,
 }
